friend_finder/requirements.txt
flask
numpy
python-dotenv
requests
